2007-05-24 12:01 UTC-0500 Ron Pinkas <ron/at/xharbour.com>   * include/hbcomp.h   * include/hbinit.h   * source/compiler/genc.c   * source/compiler/genjava.c   * source/compiler/genobj32.c   * source/compiler/harbour.c   * source/compiler/harbour.sly   * source/ct/atadjust.c   * source/ct/count.c   * source/ct/ctpad.c   * source/ct/justify.c   * source/ct/pos2.c   * source/ct/range.c   * source/ct/remove.c   * source/ct/replace.c   * source/odbc/odbc.c   * source/rdd/dbffpt/dbffpt1.c   * source/rtl/gtgraph.c   * source/rtl/gtwvt/gtwvt.c     * Compiler warnings

+++ xharbour/source/compiler/genjava.c
@@ -1,5 +1,5 @@
 /*
- * $Id: genjava.c,v 1.3 2002/01/27 22:30:07 ronpinkas Exp $
+ * $Id: genjava.c,v 1.4 2006/05/16 22:57:08 druzus Exp $
  */
 
 /*
@@ -94,7 +94,7 @@ void hb_compGenJava( PHB_FNAME pFileName )
    {
       hb_fputs( pSym->szName );
       hb_fputc( 0 );
-      hb_fputc( pSym->cScope );
+      hb_fputc( (BYTE) pSym->cScope );
 
       /* specify the function address if it is a defined function or a
          external called function */
